backend/pyproject.toml
[tool.poetry]
name = "aria-backend"
version = "0.1.0"
description = "ARIA — AI-powered business idea evaluation platform"
authors = ["ARIA Team"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.11,<=3.13"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
sqlalchemy = { version = "^2.0.0", extras = ["asyncio"] }
asyncpg = "^0.29.0"
alembic = "^1.13.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.3.0"
crewai = "^0.80.0"
crewai-tools = "^0.14.0"
anthropic = "^0.34.0"
openai = "^1.45.0"
google-generativeai = "^0.8.0"
llama-index = "^0.11.0"
llama-index-vector-stores-postgres = "^0.2.0"
pgvector = "^0.2.4"
celery = { version = "^5.4.0", extras = ["redis"] }
redis = "^5.1.0"
python-multipart = "^0.0.12"
weasyprint = "^62.3"
httpx = "^0.27.0"
structlog = "^24.4.0"
slowapi = "^0.1.9"
tiktoken = "^0.7.0"
PyMuPDF = "^1.24.0"
python-docx = "^1.1.0"
jinja2 = "^3.1.0"
tavily-python = "^0.5.0"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-httpx = "^0.30.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
